PR changes of link preview
