--- ChatSX/components/InputBox/models/LinkPreviewSnapProps.ts
@@ -1,5 +1,0 @@
-import {OgTags} from './OgTags';
-
-export interface LinkPreviewSnapProps {
-  ogTags: OgTags;
-}
