+++ ChatSX/components/InputBox/models/index.tsx
@@ -1,4 +1,5 @@
 import {InputBoxProps} from './InputBoxProps';
 import {LaunchActivityProps} from './LaunchActivityProps';
+import {OgTags} from './OgTags';
 
-export {InputBoxProps, LaunchActivityProps};
+export {InputBoxProps, LaunchActivityProps, OgTags};
