+++ ChatSX/components/LinkPreview/models/index.ts
@@ -1,0 +1,4 @@
+import {LinkPreviewBoxProps} from './LinkPreviewBoxProps';
+import {LinkPreviewProps} from './LinkPreviewProps';
+
+export {LinkPreviewBoxProps, LinkPreviewProps};
