+++ ChatSX/components/linkPreviewInputBox/models/LinkPreviewInputBoxProps.ts
@@ -1,0 +1,6 @@
+import { OgTags } from "../../InputBox/models";
+
+
+export interface LinkPreviewInputBoxProps {
+  ogTags: OgTags;
+}
