+++ ChatSX/components/linkPreviewInputBox/models/index.ts
@@ -1,0 +1,3 @@
+import {LinkPreviewInputBoxProps} from './LinkPreviewInputBoxProps';
+
+export {LinkPreviewInputBoxProps};
